Remove iMisc.cpp, and move the code to System.cpp/h.
