+++ pcsx2/pcsx2core.vcxproj.filters
@@ -353,9 +353,6 @@
     <ClCompile Include="Dump.cpp">
       <Filter>System</Filter>
     </ClCompile>
-    <ClCompile Include="x86\iMisc.cpp">
-      <Filter>System</Filter>
-    </ClCompile>
     <ClCompile Include="Pcsx2Config.cpp">
       <Filter>System</Filter>
     </ClCompile>
