--- pcsx2/x86/iMisc.cpp
@@ -1,33 +1,0 @@
-/*  PCSX2 - PS2 Emulator for PCs
- *  Copyright (C) 2002-2010  PCSX2 Dev Team
- *
- *  PCSX2 is free software: you can redistribute it and/or modify it under the terms
- *  of the GNU Lesser General Public License as published by the Free Software Found-
- *  ation, either version 3 of the License, or (at your option) any later version.
- *
- *  PCSX2 is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
- *  PURPOSE.  See the GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License along with PCSX2.
- *  If not, see <http://www.gnu.org/licenses/>.
- */
-
-
-#include "PrecompiledHeader.h"
-#include "common/emitter/x86_intrin.h"
-
-SSE_MXCSR g_sseMXCSR   = {DEFAULT_sseMXCSR};
-SSE_MXCSR g_sseVUMXCSR = {DEFAULT_sseVUMXCSR};
-
-// SetCPUState -- for assignment of SSE roundmodes and clampmodes.
-//
-void SetCPUState(SSE_MXCSR sseMXCSR, SSE_MXCSR sseVUMXCSR)
-{
-	//Msgbox::Alert("SetCPUState: Config.sseMXCSR = %x; Config.sseVUMXCSR = %x \n", Config.sseMXCSR, Config.sseVUMXCSR);
-
-	g_sseMXCSR   = sseMXCSR.ApplyReserveMask();
-	g_sseVUMXCSR = sseVUMXCSR.ApplyReserveMask();
-
-	_mm_setcsr(g_sseMXCSR.bitmask);
-}
